Fixed uxrom

+++ src/mapper/uxrom.rs
@@ -12,18 +12,18 @@ use std::rc::Rc;
 
 pub struct Uxrom {
     cart: Cartridge,
-    prg_banks: u16,
-    prg_bank1: u16,
-    prg_bank2: u16,
+    prg_banks: u8,
+    prg_bank1: u8,
+    prg_bank2: u8,
 }
 
 impl Uxrom {
     pub fn load(cart: Cartridge) -> MapperRef {
-        let prg_banks = (cart.prg_rom.len() / 0x4000) as u16;
+        let prg_banks = (cart.prg_rom.len() / 0x4000) as u8;
         let uxrom = Self {
             cart,
             prg_banks,
-            prg_bank1: 0u16,
+            prg_bank1: 0u8,
             prg_bank2: prg_banks - 1,
         };
         Rc::new(RefCell::new(uxrom))
@@ -36,11 +36,11 @@ impl Memory for Uxrom {
             0x0000..=0x1FFF => self.cart.chr_rom[addr as usize],
             0x6000..=0x7FFF => self.cart.sram[(addr - 0x6000) as usize],
             0x8000..=0xBFFF => {
-                let idx = self.prg_bank1 * 0x4000 + addr - 0x8000;
+                let idx = u32::from(self.prg_bank1) * 0x4000 + u32::from(addr - 0x8000);
                 self.cart.prg_rom[idx as usize]
             }
             0xC000..=0xFFFF => {
-                let idx = self.prg_bank2 * 0x4000 + addr - 0xC000;
+                let idx = u32::from(self.prg_bank2) * 0x4000 + u32::from(addr - 0xC000);
                 self.cart.prg_rom[idx as usize]
             }
             _ => {
@@ -55,7 +55,7 @@ impl Memory for Uxrom {
             0x0000..=0x1FFF => self.cart.chr_rom[addr as usize] = val,
             0x6000..=0x7FFF => self.cart.sram[(addr - 0x6000) as usize] = val,
             0x8000..=0xFFFF => {
-                self.prg_bank1 = u16::from(val) % self.prg_banks;
+                self.prg_bank1 = val % self.prg_banks;
             }
             _ => {
                 eprintln!(
